Replaced public 'SFC_' and 'DOMAIN_' constants with methods and enums (Mac)
The 'GetSpecialFolderConstant' method uses the 'SpecialFolderConstant' enum to return a corresponding 'special folder constant' that can be used with the 'GetSpecialFolderPath' method (former 'GetSpecialFolderMac') to retrieve the path for the required folder. The 'GetSpecialFolderDomain' method uses the 'SpecialFolderDomain' enum to return a corresponding 'domain name'. Global namespace is somewhat tidier.